A start at defining an AST architecture
The idea is that we should define an internal JS object representation that is independent of the input and output formats.  In the examples/ directory, I'm working on fleshing out a complete example that shows all the features, basically so that we can say, "If it's not in here, we don't support it."  The mkd version is complete, and the xml version was generated using kramdown (plus a couple of hand-tweaks).  I expect to flesh out the js version on the plane tomorrow.